+++ processor/src/test/resources/expected/com/example/component/Arez_DisposeOnDeactivateModel.java
@@ -42,7 +42,7 @@ public final class Arez_DisposeOnDeactivateModel extends DisposeOnDeactivateMode
     }
     this.$$arezi$$_component = Arez.areNativeComponentsEnabled() ? $$arezi$$_context().createComponent( "DisposeOnDeactivateModel", $$arezi$$_id(), Arez.areNamesEnabled() ? $$arezi$$_name() : null ) : null;
     this.$$arezi$$_disposedObservable = $$arezi$$_context().createObservable( Arez.areNativeComponentsEnabled() ? this.$$arezi$$_component : null, Arez.areNamesEnabled() ? $$arezi$$_name() + ".isDisposed" : null, Arez.arePropertyIntrospectorsEnabled() ? () -> this.$$arezi$$_state >= 0 : null );
-    this.$$arezi$$_disposeOnDeactivate = $$arezi$$_context().createComputedValue( Arez.areNativeComponentsEnabled() ? this.$$arezi$$_component : null, Arez.areNamesEnabled() ? $$arezi$$_name() + ".disposeOnDeactivate" : null, () -> $$arezi$$_observe(), EqualityComparator.defaultComparator(), null, () -> dispose(), null, null );
+    this.$$arezi$$_disposeOnDeactivate = $$arezi$$_context().createComputedValue( Arez.areNativeComponentsEnabled() ? this.$$arezi$$_component : null, Arez.areNamesEnabled() ? $$arezi$$_name() + ".disposeOnDeactivate" : null, () -> $$arezi$$_observe(), EqualityComparator.defaultComparator(), null, () -> $$arezi$$_context().scheduleDispose( this ), null, null );
     if ( Arez.shouldCheckInvariants() ) {
       this.$$arezi$$_state = ComponentState.COMPONENT_CONSTRUCTED;
     }